% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{cr}
\alias{cr}
\title{A scale and center function adapted to vectors}
\usage{
cr(v_test)
}
\arguments{
\item{v_test}{a vector.}
}
\value{
a vector.
}
\description{
This function allows to scale and center a vector.
}
\examples{
cr()
}
