% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_function.R
\name{emsem_function}
\alias{emsem_function}
\title{An simultaneous EM estimation of a structural equation model and its latent factors.}
\usage{
emsem_function(Y, X1, X2, T, T1, T2, epsilon = 10^{     -3 }, nb_it = 100)
}
\arguments{
\item{Y}{numeric output matrix (or vector) which is the dependant bloc of observed variables.}

\item{X1}{numeric input matrix (or vector) which is one explanatory bloc of observed variables.}

\item{X2}{numeric input matrix (or vector) which is one explanatory bloc of observed variables.}

\item{T}{numeric input matrix (or vector) for extra-covariates linked to Y.}

\item{T1}{numeric input matrix (or vector) for extra-covariates linked to X1.}

\item{T2}{numeric input matrix (or vector) for extra-covariates linked to X2.}

\item{epsilon}{numeric accuracy for the convergence criterion.}

\item{nb_it}{numeric for give a maximal number of iteration.}
}
\value{
list giving the reconstructed latent factor values,
 the estimated model’s coefficients, the value of the stopping criterion,
 the number of iterations effected.
}
\description{
This function allows you to fit a structural equation model with one bloc of dependant observed variables
and two blocs of explanatory observed variables. 
Underlying to each bloc there is a latent factor which is simultaneously reconstructed.
}
\examples{
emsem_function()
}
\references{
\url{https://myriamtami.github.io/data/publications/Tami_thesis.pdf}
}
\author{
Myriam Tami, \email{myriam.tami@univ-grenoble-alpes.fr}
}
\keyword{factors}
\keyword{latent}
