% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{moy_fg_tilde}
\alias{moy_fg_tilde}
\title{A mean function of a product between each element from an array and the transpose from another array}
\usage{
moy_fg_tilde(Ftilde, Gtilde)
}
\arguments{
\item{Ftilde}{an array corresponding to the explanatory factors.}

\item{Gtilde}{an array corresponding to the dependent factor.}
}
\value{
a mean value.
}
\description{
This function allows to compute in the EM algorithm 
the mean on the statistical units of the products between each 
element of an explanory reconstructed factor and each element from 
a dependent reconstructed factor.
}
\examples{
moy_fg_tilde()
}
