% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{moy_T_Prim_Data}
\alias{moy_T_Prim_Data}
\title{A mean function of a product between each row of a matrice and the transpose of another matrice}
\usage{
moy_T_Prim_Data(YX, T)
}
\arguments{
\item{YX}{a matrix.}

\item{T}{a matrix.}
}
\value{
a mean value.
}
\description{
This function allows to compute in the EM algorithm 
the mean on the statistical units of the products between each 
row of a bloc of observed data and a row of an extra-covariate matrix.
}
\examples{
moy_T_Prim_Data()
}
