% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{moy_T_T_Prim}
\alias{moy_T_T_Prim}
\title{A mean function of a product between each row of a matrice and its transpose}
\usage{
moy_T_T_Prim(T, Tbis)
}
\arguments{
\item{T}{a matrix.}

\item{Tbis}{a matrix.}
}
\value{
a mean value.
}
\description{
This function allows to compute in the EM algorithm 
the mean on the statistical units of the products between each 
row of an extra-covariate matrix and a row of the same extra-covariate matrix.
}
\examples{
moy_T_T_Prim()
}
