% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{moy_TildeData}
\alias{moy_TildeData}
\title{A mean function of a product between each element from an array and the transpose of each row from a matrice
reconstructed factor from an array and another matrix}
\usage{
moy_TildeData(Ftilde, X)
}
\arguments{
\item{Ftilde}{an array.}

\item{X}{a matrix.}
}
\value{
a mean value.
}
\description{
This function allows to compute in the EM algorithm 
the mean on the statistical units of the products between each element of the reconstructed factors and 
the corresponding instance in a bloc of observed data.
}
\examples{
moy_TildeData()
}
