% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{prod_tab_mat}
\alias{prod_tab_mat}
\title{A product function of a matrix from an array and another matrix}
\usage{
prod_tab_mat(DD, L)
}
\arguments{
\item{DD}{a matrix.}

\item{L}{an array.}
}
\value{
an array composed by each computed matrices.
}
\description{
This function allows to compute in the EM algorithm the
product of each matrix from an array and another matrix.
}
\examples{
prod_tab_mat()
}
