% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{prod_tab_rowmat}
\alias{prod_tab_rowmat}
\title{A product function of each element from an array and each row of a matrix}
\usage{
prod_tab_rowmat(mat, L)
}
\arguments{
\item{mat}{a matrix.}

\item{L}{an array.}
}
\value{
an array.
}
\description{
This function allows to compute in the EM algorithm 
the products between each matrix from an array and 
each row of a matrix.
}
\examples{
prod_tab_rowmat()
}
