% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsem_needed_functions.R
\name{trans_tab}
\alias{trans_tab}
\title{A product function of each matrix from an array and its transpose}
\usage{
trans_tab(L)
}
\arguments{
\item{L}{an array.}
}
\value{
an array composed by each computed matrices.
}
\description{
This function allows to compute in the EM algorithm the
product of each matrix from an array and its transpose.
}
\examples{
trans_tab()
}
